First version of the Hugo powered website
This is the initial version of the website.
Content is roughly equivalent to the content as it was previously
on the nanoc version. The major differences are:

1) There is no blog archive. Hugo can't do this out of the box.
2) The tag cloud does not vary the text size based on the tags
popularity. Again Hugo can't do this out of the box.

The existing cgi-bin perl scripts
=================================
These have been corrected. The originals - at least on Owens HD (the sever
may have been different) actually had invalid syntax! This has now been
fixed.
In addition the minumum set of perl dependancies has been worked out
and copied into the cgi-bin directory.

Deployment Instructions
=======================
Run hugo to generate the site:
% hugo --theme="kulawe-blue"
Copy the entire cgi-bin dir into the newly generated public dir
% cp -r cgi-bin ./public/
Copy both the .htaccess and the robots.txt into the public dir
% cp .htaccess ./public/
% cp robots.txt ./public/
Now ftp the entire public dir to the server

Content live as of 26/08/14